Changed the logic for punishment and Rolling Latency
There was a bug that was using the time from the server to subtract from the accumulatedeltatime, since that delta time is coming from the client this will cause a conflict.  Instead we work around this by just cleaning up the accumulatedDeltaTime if more than 2 seconds has passed in the last input. We can safely assume that the player stopped if IsRunning is false and timeSinceLastCommand is greater than 2 seconds.
So after 30 seconds we give the player back the ability to walk at normal speed if they have not moved and they are not cheating again.

+++ KitchenChaos/Assets/Scripts/Player/PlayerProperties.cs
@@ -19,22 +19,22 @@ public class PlayerProperties : NetworkBehaviour
     public float PlayerHeight { get => _playerHeight.Value; }
 
     [ServerRpc(RequireOwnership = false)]
-    public void SetMovementSpeedServerRpc(int movementSpeed)
+    public void SetMovementSpeedServerRpc(float movementSpeed)
     {
         this._movementSpeed.Value = movementSpeed;
     }
     [ServerRpc(RequireOwnership = false)]
-    public void SetRotationSpeedServerRpc(int rotationSpeed)
+    public void SetRotationSpeedServerRpc(float rotationSpeed)
     {
         this._rotationSpeed.Value = rotationSpeed;
     }
     [ServerRpc(RequireOwnership = false)]
-    public void SetPlayerRadiusServerRpc(int playerRadius)
+    public void SetPlayerRadiusServerRpc(float playerRadius)
     {
         this._playerRadius.Value = playerRadius;
     }
     [ServerRpc(RequireOwnership = false)]
-    public void SetPlayerHeightServerRpc(int playerHeight)
+    public void SetPlayerHeightServerRpc(float playerHeight)
     {
         this._playerHeight.Value = playerHeight;
     }
